PXP-514: [CLI] Rename 'dev config-synthesizer' to 'dev config pull' (#87)
