+++ src/commands/dev/config/mod.rs
@@ -1,11 +1,14 @@
 mod diff;
+mod pull;
 mod push;
 mod utils;
 
 use crate::error::CliError;
 use clap::{Args, Subcommand};
 use diff::handle_config_diff;
+use pull::handle_config_pull;
 use push::handle_config_push;
+use std::path::PathBuf;
 
 #[derive(Debug, Args)]
 pub struct Config {
@@ -19,6 +22,12 @@ pub enum ConfigSubcommand {
     Push,
     /// Show changes between the local configuration and the Bunker.
     Diff,
+    /// Pull the development config file from Bunker.
+    Pull {
+        /// The path to the project
+        #[arg(default_value = ".")]
+        path: PathBuf,
+    },
 }
 
 impl Config {
@@ -26,6 +35,7 @@ impl Config {
         match &self.subcommand {
             ConfigSubcommand::Push => handle_config_push().await,
             ConfigSubcommand::Diff => handle_config_diff().await,
+            ConfigSubcommand::Pull { path } => handle_config_pull(path).await,
         }
     }
 }
