+++ src/commands/dev/mod.rs
@@ -1,11 +1,9 @@
 mod config;
 mod config_lint;
-mod config_synthesizer;
 
 use crate::error::CliError;
 use clap::{Args, Subcommand};
 use config_lint::handle_config_lint;
-use config_synthesizer::handle_config_synthesizer;
 use std::path::PathBuf;
 
 #[derive(Debug, Args)]
@@ -22,12 +20,6 @@ pub enum DevSubcommand {
         #[arg(default_value = ".")]
         path: PathBuf,
     },
-    /// Synthesize the development config with secrets file from Bunker.
-    ConfigSynthesizer {
-        /// The path to the project
-        #[arg(default_value = ".")]
-        path: PathBuf,
-    },
     /// This command group contains the commands to interact with the config secrets with bunker.
     Config(config::Config),
 }
@@ -36,7 +28,6 @@ impl Dev {
     pub async fn handle_command(&self) -> Result<bool, CliError> {
         match &self.subcommand {
             DevSubcommand::ConfigLint { path } => handle_config_lint(path),
-            DevSubcommand::ConfigSynthesizer { path } => handle_config_synthesizer(path).await,
             DevSubcommand::Config(config) => config.handle_command().await,
         }
     }
